+++ src/main/php/util/Comparison.class.php
@@ -1,0 +1,41 @@
+<?php namespace util;
+
+use util\Objects;
+
+/**
+ * Implements object comparison via `hashCode()` and `compareTo()`
+ *
+ * @see   xp://lang.Value
+ * @test  xp://net.xp_framework.unittest.util.ComparisonTest
+ */
+trait Comparison {
+
+  /**
+   * Calculates a hash code based on type and members
+   *
+   * @return string
+   */
+  public function hashCode() {
+    $s= self::class;
+    foreach ((array)$this as $val) {
+      $s.= '|'.Objects::hashOf($val);
+    }
+    return $s;
+  }
+
+  /**
+   * Performs member-wise comparison
+   *
+   * @param  var $value
+   * @return int
+   */
+  public function compareTo($value) {
+    if ($value instanceof self) {
+      foreach ((array)$this as $key => $val) {
+        if (0 !== $r= Objects::compare($val, $value->{$key})) return $r;
+      }
+      return 0;
+    }
+    return 1;
+  }
+}